Module punkt::params

Contains traits for configuring all tokenizers, and the trainer. Also contains default parameters for tokenizers, and the trainer.

Structs

Default
Default settings for all tokenizers, and trainers.

Traits

DefaultCharacterDefinitions
Mixin that will give the default implementations for DefinesSentenceEndings, DefinesInternalPunctuation, DefinesNonWordCharacter, DefinesEndingPunctuation, and DefinesNonPrefixCharacters.

DefinesInternalPunctuation
Defines a set of punctuation that can occur within a word.

DefinesNonPrefixCharacters
Defines a set of a characters that can not start a word.

DefinesNonWordCharacters
Defines a set of characters that can not occur inside of a word.

DefinesPunctuation
Defines punctuation that can occur within a sentence.

DefinesSentenceEndings
Defines a set of punctuation that can end a sentence.

TrainerParameters
Configurable parameters for a trainer.

Type Definitions

Set
Type for character sets.