Trait punkt::params::TrainerParameters

pub trait TrainerParameters: DefinesSentenceEndings + DefinesInternalPunctuation {
    fn abbrev_lower_bound() -> f64;
    fn abbrev_upper_bound() -> f64;
    fn ignore_abbrev_penalty() -> bool;
    fn collocation_lower_bound() -> f64;
    fn sentence_starter_lower_bound() -> f64;
    fn include_all_collocations() -> bool;
    fn include_abbrev_collocations() -> bool;
    fn collocation_frequency_lower_bound() -> f64;
}

Configurable parameters for a trainer.

Required Methods

fn abbrev_lower_bound() -> f64

Lower bound score for a token to be considered an abbreviation.

fn abbrev_upper_bound() -> f64

Upper bound score for a token to be considered an abbreviation.

fn ignore_abbrev_penalty() -> bool

Disables the abbreviation penalty which exponentially penalizes occurances of words without a trailing period.

fn collocation_lower_bound() -> f64

Lower bound score for two tokens to be considered a collocation

fn sentence_starter_lower_bound() -> f64

Lower bound score for a token to be considered a sentence starter.

fn include_all_collocations() -> bool

Include all pairs where the first token ends with a period.

fn include_abbrev_collocations() -> bool

Include all pairs where the first is an abbreviation. Overridden by include_all_collocations.

fn collocation_frequency_lower_bound() -> f64

Minimum number of times a bigram appears in order to be considered a collocation.