Struct punkt::TrainingData

pub struct TrainingData<'a> {
    // some fields omitted
}

Stores data that was obtained during training.

Examples

Precompiled data can be loaded via a language specific constructor.

let eng_data = TrainingData::english();
let ger_data = TrainingData::german();

assert!(eng_data.contains_abbrev("va"));
assert!(ger_data.contains_abbrev("crz"));

Methods

impl<'a> TrainingData<'a>

fn new() -> TrainingData<'a>

Creates a new, empty data object.

fn contains_abbrev(&self, tok: &str) -> bool

Check if a token is considered to be an abbreviation.

fn contains_sentence_starter(&self, tok: &str) -> bool

Check if a token is considered to be a token that commonly starts a sentence.

fn contains_collocation(&self, left: &str, right: &str) -> bool

Checks if a pair of words are commonly known to appear together.

fn get_orthographic_context(&self, tok: &str) -> u8

Gets the orthographic context for a token. Returns 0 if the token was not yet encountered.

impl<'a> TrainingData<'a>

fn czech() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn danish() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn dutch() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn english() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn estonian() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn finnish() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn french() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn german() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn greek() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn italian() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn norwegian() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn polish() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn portuguese() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn slovene() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn spanish() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn swedish() -> TrainingData<'a>

impl<'a> TrainingData<'a>

fn turkish() -> TrainingData<'a>